Added getWeatherData() and some other minor changes to WeatherStation

+++ src/main/java/com/nayeel/imtiaz/patterns/observer/weather/v2/WeatherData.java
@@ -1,3 +1,3 @@
 package com.nayeel.imtiaz.patterns.observer.weather.v2;
 
-public record WeatherData(float temperature, float humidity, float pressure) { }
+public record WeatherData(double temperature, double humidity, double pressure) { }
